requirements.txt
dnspython==2.6.1
python-dateutil==2.9.0.post0
python-whois==0.9.4
six==1.16.0
